This class is just a test, not essential.
This is a first attempt to draw cell boxes above the image, including text and a semi-transparent background. 
The Canvas is a Tkinter class, so theoretically, it does not require OpenCV to create the boxes (although this code uses it, but it should work even without it). 
I still need some time to figure out how to draw all the cells in a suitable manner.
 The stipple parameter offers the possibility to change the transparency percentage from 12% to 25% to 50% and above. In this example, I used 25%, which I believe is a good compromise.