Update to the canvas table
In this commit I uploaded the testFileForRunTime file. It's now possible to have more control of the canvas.
Since the canvas will cover the whole image, width and height of each cell are calculated depending on the number of columns and row.
Rows and columns number are retrieved from a json.

I need some extra time to implement it on AdaptiveGUI, right now it just print a table of canvas with the text "test" for every cell.